Refactor common utilities to resolve matrix coefficients [breaking]
This is an attempt to simplify the workflow, by avoiding having python
calling lua, calling python, calling lua, etc..

The motivation is also to work around this bug:
https://github.com/scoder/lupa/issues/243

With such purpose in mind, I now use Lua for the generation of the
assignments of the matrix coefficients, rather than Python.

This change breaks the existing code generation backends.

Also, the code now requires the language-specifics to have an
explicit function to turn a sympy expression into text.
This is used instead of the naive call to Sympy's __str__,
to give the backend more chances for customization.

For example, powers in Sympy (Python) are written with '**',
which is not a valid operator in other languages like C.
Therefore, simply using __str__ would cause invalid code to
be generated.