Fix bug when reifying the model constants
I was replacing the float constants in one trasform at the time,
which cannot keep track of how many constants are there in the
whole model. This resulted in different constants with the same
name, because the counting (which determines the name) would restart
from 0 for every transform (e.g. c0, c1, c0, c1, c2, c0, ...
instead of c0, c1, c2, c3, c4, ...).

I changed the function to take a list of transforms rather than a
single one, so the counting is "global".