WIP - Fix feature "do not generate assignments for the 0 coefficients"
Do not check whether the value is 0, but convert it
to string and check if it is "0.0" or "0", because
the actual runtime value is possibly a
sympy.core.numbers.Zero.

Do not return nil for the 0 coefficients but simply
go through, otherwise iteration will stop even though
there are more coefficients to go through.

TODO: make the "skipping zeros" an option