Harmonize usage of the backend config and of the output directory [breaking]
Change the command line option for the backend configuration, from
-cc to -b, to respect the convention of one letter.

Be consistent with the naming of the backend configuration path (I was
using three different names, now it is 'backend_cfg' everywhere)

Remove the output directory from the configuration dictionary, and
just pass it explicitly to the code generation function, which is,
correctly, the only place where it is needed.
This simplifies the code, because now the configuration dictionary
is not needed at all, as the path to the backend config file is
the only thing that needs to be passed.

Adapted modules of the cpp backend.