Use a struct initializer list for the parameters [octave]
Use the syntax "struct(<field>,<value>,...)" for the parameters
inside a transform class (in the generated code).

That is useful to set initial values (the default from the
model or 0) at the same time as declaring the field, which
spares the extra lines of generated code that would set
initial values.

Also, with that syntax the generated code is compatible
with MATLAB code generation.

I refactored the code generator by moving outside of the
template the loop for the statements that update the parameters,
mainly because the same loop can be used for two things: the
struct initializer list, and the actual statements of the
updateParameter method.