Simplify configuration [octave]
Remove the python Configurator to keep only the lua
configuration table, as the first was not adding much.

Change the Generator constructor to expect directly
the lua configuration table.

As a side effect, the Generator becomes more flexible,
because clients can build the configuration table in
any way and are not constrained to use file paths.

The functionality of merging the default lua configuration
with the optional user's one, was moved from the Configurator
class (removed), directly to the backend module, exposed as
a dedicated function.