Fix defect in generated header in the case of templated code [cpp]
Added 'this->' in a call to a base class method, to avoid
compilation errors in the case of templated code. The error
is due to the unqualified method that is not looked up
in the base class because it is a "dependent" name.

Now 'this->' is generated regardless of whether the generated
code is templated on the scalar or not, because it is easier.