Simplify and fix the generated testing code [cpp]
Do not generate an additional header and source, all the
code to test a single transform is conveniently put in
the source with the corresponding main. This is much
simpler to read and neater. It also simplifies the
code generators.

The simplification also fixed a nasty bug manifesting with
templated code. It was an issue with the initialization order
of the model constants and the transforms objects used
in the tests.

The tests now target the individual transforms, though,
the container class is not tested anymore.

Also changed an implementation detail, use an array instead
of a matrix to read the values from the datasets. In the
generic testing code, operator[] is expected on the
given container.