Refactor and fix generation of code for the constants [cpp]
Use a dedicated file for the template and the generator.
The code is still non trivial but much much simpler then
before.

Tested with all the four combinations of configuration
stemming from the boolean flags about templates and constexpr.

Usage of constexpr for the constants is now the default,
because it is the best implementation and it is not
suitable only for non-literal scalar types.